+++ util-stats/src/main/scala/com/twitter/finagle/stats/StatsReceiver.scala
@@ -32,14 +32,6 @@ object StatsReceiver {
 object StatsReceivers {
 
   /**
-   * Java compatible version of [[StatsReceiver.counter]].
-   */
-  @varargs
-  @deprecated("Use vararg StatsReceiver.counter() instead", "2017-6-16")
-  def counter(statsReceiver: StatsReceiver, name: String*): Counter =
-    statsReceiver.counter(name: _*)
-
-  /**
    * Java compatible version of [[StatsReceiver.addGauge]].
    */
   @varargs
@@ -55,13 +47,6 @@ object StatsReceivers {
   @varargs
   def provideGauge(statsReceiver: StatsReceiver, callable: Callable[JFloat], name: String*): Unit =
     statsReceiver.provideGauge(name: _*)(callable.call())
-
-  /**
-   * Java compatible version of [[StatsReceiver.stat]].
-   */
-  @varargs
-  @deprecated("Use vararg StatsReceiver.stat() instead", "2017-6-16")
-  def stat(statsReceiver: StatsReceiver, name: String*): Stat = statsReceiver.stat(name: _*)
 }
 
 /**
@@ -126,15 +111,6 @@ trait StatsReceiver {
     counter(schema.metricBuilder.verbosity, schema.metricBuilder.name: _*)
 
   /**
-   * Get a [[Counter counter]] with the given `name`.
-   *
-   * This method is a convenience for Java programs, but is no longer needed because
-   * [[StatsReceivers.counter]] is usable from java.
-   */
-  @deprecated("Use vararg counter() instead", "2017-6-16")
-  def counter0(name: String): Counter = counter(name)
-
-  /**
    * Get a [[Stat stat]] with the given name.
    */
   @varargs
@@ -155,14 +131,6 @@ trait StatsReceiver {
    */
   private[stats] def stat(schema: HistogramSchema): Stat =
     stat(schema.metricBuilder.verbosity, schema.metricBuilder.name: _*)
-
-  /**
-   * Get a [[Stat stat]] with the given name. This method is a convenience for Java
-   * programs, but is no longer needed because [[StatsReceivers.counter]] is
-   * usable from java.
-   */
-  @deprecated("Use vararg stat() instead", "2017-6-16")
-  def stat0(name: String): Stat = stat(name)
 
   /**
    * Register a function `f` as a [[Gauge gauge]] with the given name that has
@@ -329,6 +297,11 @@ trait StatsReceiver {
         override def scope(namespace: String): StatsReceiver = self.scope(namespace).scope(suffix)
       }
   }
+
+  // These two are needed to accommodate Zookeper's circular dependency.
+  // We'll remove them once zk is upgraded: CSL-4710.
+  private[stats] def counter0(name: String): Counter = counter(name)
+  private[stats] def stat0(name: String): Stat = stat(name)
 }
 
 abstract class AbstractStatsReceiver extends StatsReceiver {