+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Breaking API Changes
 * util-core: Moved `c.t.io.Reader.chunked` to `c.t.io.BufReader.chunked`, and `Reader.framed` to
   `BufReader.framed`. ``PHAB_ID=D392198``
 
+* util-stats: Removed deprecated methods `stat0` and `counter0` from `StatsReceiver`. ``PHAB_ID=D393063``
+
 19.10.0
 -------